requirements.txt
fastapi
uvicorn
pydantic
numpy
Pillow
insightface
onnxruntime
python-multipart
faiss-cpu
psycopg2-binary
python-dotenv